moving migration data to different file
